Add platform-specific PwaService implementations
