+++ lib/web/pwa/pwa_service_base.dart
@@ -1,0 +1,6 @@
+import 'package:flutter/material.dart';
+
+abstract class PwaService {
+  ValueNotifier<bool> get installAvailable;
+  Future<bool> promptInstall();
+}
